requirements.txt
anthropic==0.95.0
fastapi==0.115.0
uvicorn==0.32.0
jinja2==3.1.6
python-dotenv==1.0.0
python-multipart==0.0.12
requests==2.33.1  # Anthropic SDKの依存。2.32.0はCVE-2024-35195でyankされたため明示固定
pytest>=9.0.0     # ユニットテスト実行用
